Move `PublisherException` to notification-publisher module
